Implement client-managed requests as resolving #9
This commit fleshes out the design for a system that enforces rate limits and makes requests on a per-user basis. The code currently works, but needs several rounds of refactoring.

Also included in this commit are slot-based classes for memory compaction. As I intend to use this library for long-lived server processes, space matters.

I've also made a few passes at the streaming modules during the course of development. That code is weakly-tested though.